dataset/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plan9-dataset"
version = "0.1.0"
description = "Dataset preparation tooling for Plan 9 fine-tuning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "gitpython>=3.1",
    "rich>=13.0",
    "pexpect>=4.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
training = [
    "torch>=2.0",
    "transformers>=4.40",
    "trl>=0.8",
    "datasets>=2.0",
    "accelerate>=0.20",
    "peft>=0.10",
    "bitsandbytes>=0.43",
    "evaluate>=0.4",
]
huggingface = [
    "datasets>=2.0",
    "huggingface-hub>=0.20",
]
server = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
]
all = [
    "plan9-dataset[dev,training,huggingface,server]",
]

[project.scripts]
plan9-dataset = "plan9_dataset.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["plan9_dataset"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
